Simplify the way we create a link UI control in the offcanvas editor
